Update Youtube Gallery 4.4.0
Update Youtube Gallery 3.5.9 > 4.4.0

+++ administrator/components/com_youtubegallery/models/linksform.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * YoutubeGallery Joomla! 3.0 Native Component
- * @version 3.5.9
- * @author DesignCompass corp< <support@joomlaboat.com>
+ * YoutubeGallery Joomla! Native Component
+ * @version 4.4.0
+ * @author Ivan Komlev< <support@joomlaboat.com>
  * @link http://www.joomlaboat.com
  * @GNU General Public License
  **/
@@ -87,7 +87,7 @@ class YoutubeGalleryModelLinksForm extends JModelAdmin
         }
 		
 
-	function RefreshPlayist($cids)
+	function RefreshPlayist($cids,$update_videolist=true)
 	{
 		$where=array();
 				
@@ -117,25 +117,32 @@ class YoutubeGalleryModelLinksForm extends JModelAdmin
 				
 		$misc=new YouTubeGalleryMisc;
 		
+		$getinfomethod=YouTubeGalleryMisc::getSettingValue('getinfomethod');
 		
 		foreach($linksform_rows as $linksform_row)
 		{
 			
 			$misc->videolist_row = $linksform_row;
-			$misc->update_cache_table($linksform_row); 
+			$misc->update_cache_table($linksform_row,$update_videolist); //false - refresh
 				
-			$query='UPDATE #__youtubegallery_videolists SET `lastplaylistupdate`="'.date( 'Y-m-d H:i:s').'" WHERE `id`='.$linksform_row->id;
-			$db->setQuery($query);
-			if (!$db->query())    die( $db->stderr());
+			if(!$update_videolist)
+			{
+				$query='UPDATE #__youtubegallery_videolists SET `lastplaylistupdate`="'.date( 'Y-m-d H:i:s').'" WHERE `id`='.$linksform_row->id;
+				$db->setQuery($query);
+				if (!$db->query())    die( $db->stderr());
 						
-			//Clear Update Info for each video in this gallery
-			$query='UPDATE #__youtubegallery_videos SET `lastupdate`="0000-00-00 00:00:00" WHERE `isvideo` AND `listid`='.$linksform_row->id;
-			$db->setQuery($query);
-			if (!$db->query())    die( $db->stderr());
-	
+
+				if($getinfomethod=='js' or $getinfomethod=='jsmanual')
+					$query='UPDATE #__youtubegallery_videos SET `lastupdate`="0000-00-00 00:00:00", `rawdata`="*youtubegallery_request*" WHERE `isvideo` AND `listid`='.$linksform_row->id;
+				else
+					$query='UPDATE #__youtubegallery_videos SET `lastupdate`="0000-00-00 00:00:00", `rawdata`="" WHERE `isvideo` AND `listid`='.$linksform_row->id;
+			
+				$db->setQuery($query);
+				if (!$db->query())    die( $db->stderr());
+			}	
 		}
 				
-				return true;
+		return true;
 	}
         
 
@@ -171,19 +178,7 @@ class YoutubeGalleryModelLinksForm extends JModelAdmin
                 echo 'Cannot check.';
         		return false;
         	}
-				
-				
-				if($linksform_row->id!=0)
-				{
-						require_once(JPATH_SITE.DS.'components'.DS.'com_youtubegallery'.DS.'includes'.DS.'misc.php');
-						$misc=new YouTubeGalleryMisc;
-						$misc->videolist_row = $linksform_row;
-						$misc->update_cache_table($linksform_row); 
-						$linksform_row->lastplaylistupdate =date( 'Y-m-d H:i:s');
-				}
-				
-						
-						
+
         	// Store
         	if (!$linksform_row->store())
         	{
@@ -193,6 +188,12 @@ class YoutubeGalleryModelLinksForm extends JModelAdmin
 				';
         		return false;
         	}
+		
+		require_once(JPATH_SITE.DS.'components'.DS.'com_youtubegallery'.DS.'includes'.DS.'misc.php');
+		$misc=new YouTubeGalleryMisc;
+		$misc->videolist_row = $linksform_row;
+		$misc->update_cache_table($linksform_row,false); 
+		$linksform_row->lastplaylistupdate =date( 'Y-m-d H:i:s');
 				
         	$this->id=$linksform_row->id;
 			
@@ -200,23 +201,24 @@ class YoutubeGalleryModelLinksForm extends JModelAdmin
         }
         
     		
-		function deleteVideoList($cids)
+	function deleteVideoList($cids)
         {
 
         	$linksform_row = $this->getTable('videolists');
 
-            $db = JFactory::getDBO();
+		$db = JFactory::getDBO();
             
         	if (count( $cids ))
         	{
         		foreach($cids as $cid)
         		{
-						
+				$query='DELETE FROM `#__youtubegallery_videos` WHERE `listid`='.(int)$cid;
+			
+				$db->setQuery($query);
+				if (!$db->query())    die( $db->stderr());
 				
 				if (!$linksform_row->delete( $cid ))
-				{
 					return false;
-				}
 			}
         	}
